qemu: Switch to use qemu for running GCC tests
This commit is a collection of fixes.  The overall goal is to include
qemu in the list of packages that are checked out and built, we then
make qemu available for running the GCC tests.  This means that we can
now test GCC even if we are building a binutils-gdb branch that does
not include the simulator.

The clone script gets a spring clean, most packages are now cloned
from their upstream sources rather than the embecosm specific
versions, RISC-V is now widely supported upstream.

I've also taken the opportunity to remove the need for a separate
binutils and gdb checkout, RISC-V support is now upstream, so we no
longer need this split format.

Files Changed:

	* build-tools.sh: Build qemu, install run qemu run script wrapper,
	clean up paths to source code, other minor cleanup.
	* clone-all.sh: Clone from upstream repositories, drop -dev flag.
	* run-scripts/riscv32-unknown-elf-run: New file.
	* run-scripts/riscv64-unknown-elf-run: New file.