Allow build script to skip QEMU is the source is not available
The nightly GDB tests don't check out QEMU.  In order to allow the
same build script to be used this commit makes it so that the QEMU
build will be skipped if the source is not available.

The run-scripts, which wrap around QEMU are also only installed if the
QEMU source is available.

Files Changed:

	* build-tools.sh: Make building QEMU optional.